net/wireguard: deal with kmod service status; closes #3273
Shift the name from wireguard-go to wireguard in this case.
HA between go and kmod is probably not going to work fully
as the service name shifts, but eventually go should be gone.

Since wireguard rc.d script misuses "status" we don't know
what the status is and since the kmod has no "process" either
we have nothing to go on so the stop and start actions cannot
be cleanly represented and are instead replaced by a single
restart.

If you want to stop wireguard from the GUI it needs to be
disabled in its settings.